Working indicators on base station. Show sampling progress and afterwards indicates the number of observed satellites. Added also install script for startup service.

+++ piksi_status_indicator/install/install_indicator_autostart.sh
@@ -2,11 +2,19 @@
 echo "Do you wish to configure led status indicator autostart? [y or Y to accept]"
 read configure_base_autostart
 if [[ $configure_base_autostart == "Y" || $configure_base_autostart == "y" ]]; then
-  echo "Configuring /etc/systemd/system/led_status_indicator.service"
-  sudo rm /etc/systemd/system/led_status_indicator.service
+  SERVICE_FILE=/etc/systemd/system/led_status_indicator.service
+  if [[ -f "$SERVICE_FILE" ]]; then
+    echo "Reconfiguring /etc/systemd/system/led_status_indicator.service"
+    sudo rm /etc/systemd/system/led_status_indicator.service
+  else
+    echo "Creating new service in /etc/systemd/system/led_status_indicator.service"
+  fi
+  
   sudo sh -c "tee -a /etc/systemd/system/led_status_indicator.service << END
+
 [Unit]
 Description=Start led status indicator and rosserial connection with arduino board automatically on startup.
+After=base_station.service
 
 [Service]
 Type=forking
@@ -17,7 +25,14 @@ User=$USER
 [Install]
 WantedBy=multi-user.target
 END"
-fi
 
 sudo systemctl daemon-reload
 sudo systemctl enable led_status_indicator
+fi
+
+echo "Would you like to set up an udev rule for your arduino? [y or Y to accept]"
+read configure_udev_rules
+if [[ $configure_udev_rules == "Y" || $configure_udev_rules == "y" ]]; then
+  echo "Creating new udev rule. The arduino will then be assigned the device path: /dev/arduino"
+  sudo cp /home/$USER/catkin_ws/src/ethz_piksi_ros/piksi_status_indicator/install/98-arduino.rules /etc/udev/rules.d/.
+fi
